requirements.txt
feedparser==6.0.11
google-generativeai==0.7.2
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
